refactor: Move calc_{protection_msg_,}sort_timestamp() to `impl ChatId`
